refactor(tokens): move the class renderer out of the checkpoint command
session tokens needs the same table, and a function two commands call
should not live in one command's file. Verbatim move plus one rename
across six references; the renderer's six existing tests pass with no
assertion changed.
